Изменение : Внутренние сплиттеры - начал подготовку по вынесению размера кеша для файловых операций(по дефолту 64Kb), а так же мин/мах размер очереди пакетов(сейчас 256Kb/128Mb соответственно) во внешние настройки, чтобы имим можно было управлять из самой программы. Шаг 1 - избавился от констант, вынес в переменные, считывание из настроек. Изменение : AMR Source - на самом деле именно Splitter.
git-svn-id: svn://mpc-next.ru/mpc-be/trunk@2163 d191f3aa-8897-45be-b9a4-56e5fb472b67

+++ src/filters/parser/MatroskaSplitter/MatroskaSplitter.h
@@ -35,13 +35,13 @@
 class MatroskaPacket : public Packet
 {
 protected:
-	int GetDataSize() {
+	size_t GetDataSize() {
 		size_t size = 0;
 		POSITION pos = bg->Block.BlockData.GetHeadPosition();
 		while (pos) {
 			size += bg->Block.BlockData.GetNext(pos)->GetCount();
 		}
-		return (int)size;
+		return size;
 	}
 public:
 	CAutoPtr<MatroskaReader::BlockGroup> bg;
